[FIX] sale_renting: rental_configurator: use the good JS model
`RentalConfigurator` used the new `RelationalModel` for the form wizard
but this model does not support very well the `onChange`, thus causing
some synchronization problem.

Right now the model used for the form view is the `RelationalModel`
which is imported from `@web/views/relational_model`. This last one
support correctly `onChange`

This commit fixes this bug by replacing the model by the good one.

Steps to reproduce the bug:
- Install sale_renting and stock
- Go to Renting and create a new rental orders
- In the order line, put a printer in the product
- A wizard is opened, in the Dates, put a day in the future so that
  printers are available.
- Add two printers in the "Serial numbers" fields, you can see that the
  quantity is not well synchronized

closes odoo/enterprise#37627

Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>