[FIX] account_reports: debug column should not appear if analytic
- Activate Analytic Accounting
- Open P&L
- Group By an Analytic Account
=> The debug icon is doubled

We even shouldn't see the debug column as we have several columns,
so the numbers it shows are pretty useless.

The decision has been made to separate the generation of the column
headers from the columns, so we can do an init_options between them.
We create the headers in a `_init_options_column_headers` that is
done before the columns, and we override these headers in analytic,
so the columns are created following these new headers.

closes odoo/enterprise#37765

Signed-off-by: John Laterre (jol) <jol@odoo.com>