[IMP] sale_planning: add domain depending on role in plan dialog
Before this commit plan dialog in planning gantt view show all
the slots for any resources and it may cause problems to user
because user has to find slots that are resource related.

This commit add a domain in plan dialog to see slots that
have role in resource roles this way user only see slots
that are actually gonna usefull for resource.

task-2888248

closes odoo/enterprise#29628

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>