[IMP] helpdesk: improve UI and UX
Before this commit, the users seemed confused between the
`partner_id` and `partner_name` fields during the onboarding.
The `partner_name` is normally only useful when receiving tickets
from an alias or a contact form and so it prevents
the contact database from being bloated with spam since the
system will search an existing partner with name and email set
instead of directly creating a partner during the creation of the
ticket.

This commit removes the `partner_name` and `partner_email` fields
from form view of `helpdesk.ticket` model since those fields are
not really useful during the ticket creation with the form view.
This commit also prevents to click on partner_id field to go to
the form view of `res.partner` in the list view of `helpdesk.ticket`
to avoid having too many clickable elements inside of a clickable
element.

task- 3186602

closes odoo/enterprise#37671

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>