[FIX] account_asset: autorize misc entries for original_move_line_ids
We can have a salary cost (so misc entries) capitalized.
So, we need to be able to create assets with these.
So, the condition on the domain on `original_move_line_ids` has to
be lighter, to allow miscellaneous entries and purchase receipts.

We now have to prevent the entries of other assets to appear here.

task-3178241

closes odoo/enterprise#37694

X-original-commit: c154057c3699ad3f14e0b694d4436d44f3a3a76f
Signed-off-by: William André (wan) <wan@odoo.com>