[FIX] l10n_cl_edi_stock: adjust the readonly attr for CL
Fix the readonly field for Peru.

Steps:
Install l10n_pe_edi_stock
Install l10n_cl_edi_stock
upload CAF, signature and all the things needed to use electronic invoicing for chile
Create a sale, confirm and create a picking. Validate.
Press create delivery guide.

The system claims that there is no a sequence for the chilean delivery guide. Then edit, and put in l10n_latam_document the first number. Confirm it.
Even when the field seems editable, you will get a message claiming that you must define a document number.
What is happening is that the field is not saved.

After the PR: Odoo allows you to define the sequence for delivery guides, and the first document to issue.

closes odoo/enterprise#39251

X-original-commit: 246bca48deb7ceb0ac007fef35af5fc4bc82a209
Signed-off-by: Josse Colpaert <jco@odoo.com>