[FIX] website_knowledge: adds separator to website_knowledge template
This commit remove the condition on the separator in the website_knowledge
module to separate the added buttons by inheritance from the rest of the
share panel.

Before this commit, when an article was readonly the separator was not present.
Now, either if we are in readonly or not the separator is visible.

task-3211699

closes odoo/enterprise#37707

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>