[IMP] knowledge: debounce searching articles
This commit adds a debounce when searching for articles in:
- The "move to" popup window
- The /article command (that adds a link to an article)

These two features both implemented a select2 input that did a call to "search"
every time the user typed a single character.
We now wait for 500ms in-between key strokes to lighten the load on the server.

In addition, we slightly improved the /article domain to avoid putting an
unnecessary "name, ilike, %%" in the search domain when opening the dialog.

Task-3169282

X-original-commit: 31391cadaa0355f72284bbec2889eda7bf483786
Part-of: odoo/enterprise#39928