[IMP] knowledge: only load 50 first articles of a sub-tree
PURPOSE

This is a performance commit aiming to solve issues when the "tree panel" on the
side of the opened article tries to load a large amount of articles at once.

We introduce a "pager" that will only load the first 50 articles in a sub-tree.
A "sub-tree" is the root articles, all the articles of parent A, all the
articles of parent B, etc.

(Side-note: rest of this commit message assumes the limit to be 50).

The user can then click on a "Load more" link to load the next 50 in that
sub-tree.

TECHNICALITIES

* Main thinking

As the template rendering receives ALL articles that are going to be shown in
a single query, we cannot limit the initial query to "50". As our limit should
affect each individual sub-tree and not the entirety of displayed articles.

However, during testing, we noticed that fetching a large amount of articles is
not the performance bottleneck. The bottleneck resides in rendering the HTML of
all these articles and then sending a very large data blob to the client.

By limiting the rendering at the template level, even if probably not as
performant as reducing the amount of records fetched from the database, we
drastically reduce the tree loading time.

To give some figures, loading the root articles containing about 1.000 items
took a little bit over 10 seconds in local testing.
Only rendering the first 50 takes this number down to a few hundreds
milliseconds.

When the user clicks on "load more", we then only load exactly 50 records, as
we are loading a subset of this specific sub-tree.

* Active article special use case

As we want to always show the active article *and its ancestors*, even if it's
not within the first 50 articles of its own sub-tree, we need to include some
rather complicated code.

Indeed, we need to know if the active article, or any of its ancestor, is
present in the sub-tree we are rendering. If it's the case, and if it's not in
the first 50, then we "force" its rendering by manually including it as a 51th
element at the bottom of our result.

Upon clicking on "Load more", this "extra article" is removed as it could be
within the next 50, if not, we apply the same logic again.

We need to apply this additional active article for both:
- The first rendering where we have all articles at once
- The subsequent renderings of the "next 50 articles"
  (where values are computed on controller side)

* Example + Limitation

There is one (small) limitation to the "active article special use case".

Consider 53 root articles.
"Root Article 52" has 53 children.
The active article is "Child Article 53".

Upon first loading
- As "Root Article 52" is not within the first 50 and its an ancestor of our
  active article, we force it to show after the first 50.
- As "Child Article 53" is not within the first 50 of its sub-tree and its our
  active article, we force it to show after the first 50 children of
  "Root Article 52".

Now if the user first clicks on "Load more" to load more *children* articles.
It will show all children articles, so far so good.

But if the user then clicks on "Load more" to load more *root* articles.
It will show all root articles BUT reset what we "loaded more" on children.
This is because the sub-tree of children articles is re-evaluated when loading
more root articles, and we "loose" what was loaded before.

If these actions are taken the other way around, then everything works fine.

* Misc

This technique is used for both the backend and the "frontend" (portal page).
Some if/else are included to ensure compatibility between the backend framework
and the publicWidget methods.

We decided to merge this improvement in stable (16.0) as it prevents poor
performance and/or crashes for large amount of data.
It should have a limited impact as it does not remove parts of the initial
templates or any of the variables passed to them from the controller.

LINKS

Task-3169282

closes odoo/enterprise#39585

closes odoo/enterprise#39928

X-original-commit: dc981b1e944dde49dec2d00d2cc1058288f5a8e9
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>