[REF] web: Trigger events on the model itself
In the model, we trigger events RELATIONAL_MODEL:NEED_LOCAL_CHANGES,
RELATIONAL_MODEL:WILL_SAVE_URGENTLY and RELATIONAL_MODEL:FIELD_IS_DIRTY
on env.bus. That means that if we have two instances of the model in //
(e.g. two views, or two Record components, or a mix...), their fields would listen and react to events
triggered for another model. Trigger those events on the model instead,
as it is an EventBus.

Because the events are directly triggered on the model, we can remove
"RELATIONAL_MODEL:".

Part of task: 3179751

closes odoo/enterprise#37727

Related: odoo/odoo#113880
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>