[FIX] delivery_ups: use customer name for SoldTo
Steps to reproduce:

1. Set up UPS shipping, choose an international delivery service
2. Make an international sale order to a customer with no company
3. Validate the delivery, get error from UPS API

Before this commit Odoo used commercial company name for name field of
SoldTo. If the customer had no company assigned, it would cause an
error because the name was empty.

In this case, we can use the name of the customer (similar to shipper)
to avoid error. We also need to limit the length of this field to 35
characters.

opw-3191379

closes odoo/enterprise#38928

X-original-commit: b79df7b32001e5519a9bf0992c8a8fbc336e0e50
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Khanalizadeh Ahmad (khah) <khah@odoo.com>