[REF] analytic_enterprise: convert grid view in OWL
Before this commit, an override was made in the grid view to
change the period start and end according to fiscal year dates
defined in the current company when the range scale is `Year`.

This commit converts this override and it is now made in custom new
method called `grid_compute_year_range` and this method will be called
before the columns are generated if the range scale is in `Year`.

task-2702636