[IMP] planning: change planning sharing system
In this commit we change the way that we share the
planning slots to non-user employees and the way that
we read those slots from the front-end.
Prior to this commit, there could be asynchronisities
between the back-end and the front-end planning
which could lead to slots being visible to some employees
and not to others.

Also in this commit, we removed the "Publish" button and
its functionality from all the view of the planning app.
The button was removed as it's functionality was nearly
identical to the "Publish & Send" button, which in addition
to publishing shifts, also sends an email notification.

task-3117580

closes odoo/enterprise#37806

Related: odoo/upgrade#4406
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>