[IMP] sign: change folder structure to backend_components and common_components
Historically, sign JS codebase was always divided in backend and common folders
However, when some views were converted to OWL a new components folder was
added, therefore breaking this structure. This commit changes the sign folder
structure to use backend_components and, in the future, a common_components
folder will be added. This division makes it easier to add the files to the
correct bundle in the manifest file.

task-3071044

Part-of: odoo/enterprise#36168