[IMP] sign: convert sign.Template to OWL
This commit converts the sign.Template client action to OWL. This is the first
part of a series of tasks seeking to convert the whole sign module to OWL and
stop using jQuery and jQuery-based libraries.

Due to some specificities of the sign module, it was not possible to use OWL
inside the iframe. This is due to the fact that the content inside the iframe
is controlled by PDF.js and, since we need to add DOM our own elements in
different pages rendered by a third-party library, using OWL didn't make a lot
of sense as it was not possible to mount an OWL application on a single element
inside the iframe. Therefore, it was decided to use plain JS to handle the
rendering logic inside the iframe. In order to keep the developer experience
closer to OWL development, it was decided to use a JS class responsible for
rendering our custom elements and handling all interactions with the iframe.
This also allows us to extend the behavior of the base class as we want to
have different iframe behaviors in the frontend/backend.

Even inside the iframe class, it was possible to use renderToString to
dynamically render the sign items, so we were still able to use owl templates.
Also, it is possible to pass OWL services to the class in the constructor, so
we were able to use the rpc, orm, dialog and popover services inside the
iframe class.

On top of that, a tour was created to test the template edition functionality
and guarantee that it isn't broken.

task-3071044

closes odoo/enterprise#36168

Related: odoo/odoo#116499
Signed-off-by: Arnaud Joset <arj@odoo.com>