[FIX] knowledge: correctly rebind mousedown ev to add cover
Before this commit, and with the last version of owl [1], the
knowledge_cover_selector_tour tour failed almost all the time,
because at some point, the "Add cover" button had no mousedown
event on it. This is because the useEffect used to add the event
handler didn't correctly defined its dependencies: when the button
changes, the handler must be re-added. It has been highlighted by
the update of owl because it slightly changes the way refs are
handled, and something that was luckily working before (whereas it
shouldn't have) wasn't anymore.

[1] odoo/odoo@abaf9d0e70a480252afcf0e08f2a3c68fbbafa00

Runbot error 17846

closes odoo/enterprise#37748

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Co-authored-by: Samuel Degueldre <sad@odoo.com>