[IMP] knowledge: Move add buttons to the topbar
Purpose:
========
Move the "Add Icon", "Add Cover" and "Add Properties" buttons from the
article body to the topbar, to remove the margin between the title and
the topbar, and to make them easier to find (they now appear when hovering
the topbar). This also allows to remove the useEffects from the icon and
from the cover widgets related to these buttons, since `t-on` can now be
used on these buttons.

The "Add Properties" button is now also shown for root articles, but it
is disabled and has a tooltip explaining that the feature is not available
for root articles, so that users don't search for this button on root
articles.

This commit also fixes the properties panel not showing when hiding it and
opening another article with the same number of properties. Now, the
properties panel is always opened by default if an article has properties.

It also fixes the "last edited" and "create" dates in the option panel,
which were not updated properly: previously, these were updated when the
topbar was patched. However, since the option panel uses bootstrap, the
topbar is not patched when opening this option panel. Therefore, if you
edited an article, waited for a few seconds, and opened the options panel,
it was written that the article had been last modified 0 seconds ago. Now,
these dates are updated when opening the bootstrap dropdown instead of
when the topbar is patched.

This commit also cleans the code of the renderer and the topbar, by trying
to remove duplicate and dead code as much as possible.

Task-3193808

Part-of: odoo/enterprise#37250