[FIX] planning: handle the cases were save actually deletes
Steps:
- Create a planning shift, set a recurrence for ever.
- Go few occurences later (form via kanban) and set...
...recurrence type to "Number of Occurences",
...recurrence nmber to 1.
- Save.

Issue:
"This record doesn't exist".

Cause:
The write method will update the recurrence
so that it only contains only one occurrence.
Consequently, the record you just updated is deleted.
Then we try to fetch the record to display its form again.
Because de read returns [], _fetchRecord rejects the promise,
and we're stuck.

Fix:
In community commit, we make `FormController.saveButtonClicked` call
`Record.save` with its params.
In this commit, we override `FormController.saveButtonClicked` to
call super with the arg `throwOnError` so that, if the shift is
deleted when updating its recurrence, we will catch it and,
rather than staying on its form, go back to the previous view.

closes odoo/enterprise#37752

Related: https://github.com/odoo/odoo/pull/112335
Related: odoo/odoo#112335
Signed-off-by: Audric Onockx (auon) <auon@odoo.com>