[FIX] hr_appraisal_survey: fix prepare survey answers for multiple feedback
If applied, this commit will solve the feedback survey issue.

Before this commit:
=====================
when we ask for feedback from any recipient(s) multiple times. it will raise an
error like this 'list' object has no attribute 'filtered'. It will raise when we
select the same recipient(s) in the same appraisal form again and click on the
'Send' button.

After this commit:
=====================
It will allow work multiple times when we click on the 'Send' button and returns
the existing_answers. There is some variable issue that is also fixed in this
commit: 'answer' > 'answers'.

see - https://tinyurl.com/2fyjufzx
sentry - 3968724337

closes odoo/enterprise#37980

X-original-commit: 555c47ce0009e591c4a4cfcf41afbee4157c2ed4
Signed-off-by: Kevin Baptiste <kba@odoo.com>