[FIX] website_sale_renting: prevent the zero division error
When the duration of the product becomes Zero the server throws the Zero
Division Error, the duration should contain some value so that the ratio should
be either the calculated ratio or 1.

Ex. In the 'web_sale_renting' we can improve the
'pricing.recurrence_id.duration' as the value cannot be zero.

sentry-3932724540

closes odoo/enterprise#37821

X-original-commit: 7f85d6f6e4c6ecd52ad90491e2453df1d0adea44
Signed-off-by: Anh Thao PHAM <pta@odoo.com>