[FIX] website_sale_renting: prevent the zero division error
When the duration of the product becomes Zero the server throws the Zero
Division Error, the duration should contain some value so that the ratio should
be either the calculated ratio or 1.

Ex. In the 'web_sale_renting' we can improve the
'pricing.recurrence_id.duration' as the value cannot be zero.

sentry-3932724540

closes odoo/enterprise#37507

Signed-off-by: Anh Thao PHAM <pta@odoo.com>