[MOV] knowledge: move the abstract_behavior wysiwyg template
The template was not the owl template of the Component AbstractBehavior, so it
makes more sense to put it with other templates used in a similar purpose.

Since those templates will be pre-rendered outside the editor, the loader
inside is not necessary anymore (see next commit: fully implement oeProtected
and oeTransientContent).

Task-2821374

Part-of: odoo/enterprise#33483