[FIX] timesheet_grid: avoid to show empty section in all timesheet
In the timesheet grid view,
there is an incosistency between _apply_grid_grouped_expand and _group_expand_employee_ids.

- The first one that take care of adding the project under the employee
section if the employee has some timesheet whithin the past week
for task in that project that are active or ts not link to any task

`'|',  ('task_id.active', '=', True), ('task_id', '=', False)`

- The second one take care to show the section (the employee) if
there is some timesheet link to a project the past week

Combine with the ts automatically created when any employee take a
day off. Those timesheet are link to an inactive task in an inactive
project.

We end up with a lot of empty section because the first method ignore
those TS (they are in an inactive task) but the second one find them
and display the employee

This lead for example in a company with 2.5k employee were only 750
really timesheet to display more than 1500.

The size of the answer may lead to memory error.

We actually don't want to see empty section without project and
we don't want to see the employee that have only timesheet when they
take a day off

Solution
--------

Make both method consistent by not looking for ts in inactive task in
_group_expand_employee_ids. This help to show only the relevant project
and employee reducing the size of the answer (avoiding the memory error)
and making the page a bit more responsive.

closes odoo/enterprise#37770

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>