[FIX] l10n_ec_edi: Tax code for withhold report
The "code" used to assemble the withhold XML is not the same as "tax code"
of the tax (SRI likes to have special codes for assembling the XMLs that
differs from the "tax code").

The fix uses the "tax code" for the withhold pdf. Please note that for
profit withhold and vat withhold a different field (tax code,
and base tax code)  is used, because of local common practise.

closes odoo/enterprise#37777

Signed-off-by: Josse Colpaert <jco@odoo.com>
Co-authored-by: Andres Calle <andres.calle@trescloud.com>