[FIX] industry_fsm_stock: adapt to quant reservation
Industry fsm can choose a lot to use directly in their task and
it will add it to the linked sale order. Before if the lot was
not totaly available we reserve another ones and wait the validation
before using the right lot.

With the stock.move.line code updates, it's easier to set a lot
and reserve it. So we adapt the code here to have a reservation
that match the more possible the quantities/lot used in the
task.

closes odoo/enterprise#38207

Related: odoo/odoo#115328
Signed-off-by: William Henrotin (whe) <whe@odoo.com>