[FIX] timesheet_grid: validate the timesheets before the current date
Before this commit, when the user validates timesheets and he validates
a timesheet in the current date or in the future then the employee who
is the author of the validated timesheets will not be able to start a
new timer for the current date since the date of the new timesheet will
not be after the last validation date set on that employee.

This commit fixes the issue by validating only the timesheets before the
current date to be sure any users can start a timer and so create a
timesheets in the current date.

Issue found during the development of task-2702636

Backported in 16.0 for opw-3248046

closes odoo/enterprise#39518

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>