[FIX] l10n_de_pos_cert: keep creating TSS
Before v16, only the updated values were being given to the `write` method of the pos_config.
In v16, the pos_config settings have been moved to the general settings which then keep calling
the pos_config `write` method with all existing values. Thus it tries to send a request to the create
a new TSS which is a wrong behavior.

closes odoo/enterprise#37839

X-original-commit: d635740aa7f26f0cdd3679b8d48513eac4fffe1b
Signed-off-by: Heinz Robin (rhe) <rhe@odoo.com>
Signed-off-by: Trinh Jacky (trj) <trj@odoo.com>