[FIX] l10n_de_pos_cert: keep creating TSS
Before v16, only the updated values were being given to the `write` method of the pos_config.
In v16, the pos_config settings have been moved to the general settings which then keep calling
the pos_config `write` method with all existing values. Thus it tries to send a request to the create
a new TSS which is a wrong behavior.

closes odoo/enterprise#37820

Signed-off-by: Heinz Robin (rhe) <rhe@odoo.com>