[FIX] planning: fix allocation hours aggr. flexible hours
Current behaviour:
Putting a resource on a shift in planning with an employee that has
flexible hours, the contribution to the aggregated pill for the day
is not the same as the allocated hours on the shift.

Expected behaviour:
The allocated hours on the shift with a resource that has flexible
hours, should be the same as it's contribution the aggregate pill.

Steps to reproduce:
- Install Employee, Planning
- Set the flexible_hours on an employee
- Add it on an shift from 13h->19h (6hours allocated)
- Observe that the aggregate pill is 4 hours, because it takes the
  resource calendar into account. (13h->17h)

Reason for the problem:
There was no handling of users with flexible_hours for the
computation of the aggregate pill in the planning for the day.

Fix:
Send from the server the `flexible_hours` per resource and calculate
the contribution of with the formula `end_time - start_time` capped
by the company average hours per day.

Affected versions:
- 16.0
- saas-16.1
- master

opw-3174932

closes odoo/enterprise#37822

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>