[REF] timer,timesheet_grid: convert timer timesheet grid + timer header
This commit commits convert timer timesheet grid view in OWL ("My
Timesheets") and also timer header (component used in that grid view)
Since the timer header is converted in OWL, this component has been
added in list and kanban view to be able to start a timer in those views
as we can do it in grid view. The shortcuts to start or unlink the timer
are available in the 3 views.

task-2276491
task-2702636

Co-authored-by: Laurent Stukkens (LTU) <ltu@odoo.com>
Co-authored-by: Audric Onockx (auon) <auon@odoo.com>
Co-authored-by: Elisabeth Dickinson <edi@odoo.com>