[FIX] account_intrastat: totals line dict structure
Current behaviour: Selecting either arrival or dispatch from the type
filter on the intrastat report results in a traceback when there are any
lines on the report. This is because the totals that are passed to the
total line, which is generated when only one of these filters is
selected, are structured differently to the way in which it is
referenced.

This commit adapts the 'total_values_dict' that is passed to the
function that generates the totals line to match the structure that is
expected. In doing this, only the 'Value' column is initialised, which
means that only the total value will be displayed on the totals line,
which is fine because it's the only column that makes sense to total on
this report (total weight, supplementary units, etc are irrelevant)

A test is also updated in order to test the total line conditionally
appended at the end of the report.

closes odoo/enterprise#37830

Signed-off-by: John Laterre (jol) <jol@odoo.com>