[IMP] account_reports: trial balance readability
Increase of the minimum width for the first column of all reports. This is a revert from the following bugfix PR: https://github.com/odoo/enterprise/pull/35025
As the issue of this mentioned PR was to reduce the width of the first column of the Partner Ledger report for only PDF version, this commit specifically targets this objective. The issue with the solution provided for the mentioned PR was that the width was too small and some customers were not able to read the name of the accounts in the first column.

closes odoo/enterprise#39088

Task: 3211840
Signed-off-by: Quentin De Paoli <qdp@odoo.com>
Signed-off-by: hupo-odoo <hupo@odoo.com>