[FIX] account_followup: filter out invoices in follow-up reports while printing
Currently we can only print Invoices. So, when we try to print the follow-up
Reports containing Bank statements for Customers in Accounting,
it will raise Error as the reports contains documents other than Invoices.
Ex. Bank Statements.

So, to resolve this we filter out Invoices from the Follow-up reports
for printing.

sentry-3928914484

closes odoo/enterprise#37851

Signed-off-by: Grazioso Andrea (agr) <agr@odoo.com>