[FIX] knowledge: based_on is not set for untitled
This commit fixes a bug where if an article is Untitled, when you invite
partners to the article, all the permission inheritance is badly computed
and when modifying the permission for a member there is inconsistencies.

Before when modifying the permission for a member the modification could be
propagated to the parent of the article and none of the child articles would
have its permissions set as inherited.
Which results in inconsistencies with the original behavior.

Now when you have an article that is based on an Untitled article we correctly
set for the member the correct value for the inheritance.

Steps to Reproduce on a Runbot
==================

  * Create a tree with a depth of minimum 3
  * Invite Demo as a reader on the Root
  * Change the permission for Demo on one of the children to Write or None
  * You will see that the parent of the article you modified will have a
    different permission than before (which shouldn't happen).&

task-3220909

closes odoo/enterprise#37848

Signed-off-by: David Beguin (dbe) <dbe@odoo.com>