[FIX] web_gantt: non deterministic test
The test "position of no content help in sample mode" did wrongly
assume that the first row is always associated with the user_id 1. But
this might not be the case. Actually the rows in that tests are
constructed in the following way:

- In _populateModels, the sample server creates a given number of records
  each with a random value between 1 and 5 for the field user_id.
- in _mockReadGroup, the records with same ids are gathered to construct
  the groups that will be returned.
- in _generateRows, the gantt model generates its rows on the basis of
  the groups it has received from read_group (mocked via _mockReadGroup).

One sees that if no records is associated with user_id 1 in the first
step (that happens rarely), no row has 1 as associated resId. This was
the root cause for the test failing: all the rows had the class
o_sample_data_disabled in that situation.

Actually, that situation had probability (4/5)^16 ~ 0.028 to occur.

closes odoo/enterprise#37857

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>