[FIX] delivery_ups: prevent singleton error while multi SO Issue
Issue
=====
When the UPS Shipping connector is installed and multiple SOs are confirmed in batch,
a singleton error is raised. It is hard to reproduce the error in standard Odoo since
there's no hook for multi-SO confirmation. But when any customer tries to confirm
multiple sale orders using customizations then a singleton error will raise
because of this delivery_ups module.

In this commit, we fix the SO's action confirms method so it can be run in batch.

closes odoo/enterprise#38803

X-original-commit: a28ef18ce46e061288761930d9265746204f3b0b
Signed-off-by: Tiffany Chang <tic@odoo.com>