[FIX] knowledge: use correct cover_image_position field in frontend
Purpose:
========
Replace the nonexistent `article.cover_height` field by the correct
`article.cover_image_position` field in the frontend view.

This commit also adds a cover to an article in the frontend tour to make
sure that covers do not make knowledge crash in frontend.

It also fixes the icon position that was not overlapping the cover in
frontend anymore, and it removes the unnecessary gap between the icon
and the body in frontend

Task-3221175

closes odoo/enterprise#39862

X-original-commit: c0630bb6459f262b32183fee4f42b11db4c356a6
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>