[REF] appointment: use meaningful responses from step validation
The purpose of this commit is to use the `action_validate_step` and its
meaningful responses for step validation introduced in the COM PR to
better decide what to do thereafter and simplify the code.

Tests are included to check this behavior.

Additionally, for obvious convenience and performance reasons, we use
the "action" service to reload the current action instead of reloading
the whole app.

Task-3074112

Part-of: odoo/enterprise#37870