[FIX] appointment: float division by zero issue in appointment slot
If applied, this commit will solve division by zero issue in the appointments.

Before this commit
When the appointment duration is 00:00 Hrs and there is only one Availability
line.  It will raise an issue while previewing the appointment slots.

After this commit
Added constraints for the appointment durations as it is taking 00:00 as the
required value. It will solve the issue and stop the user while saving the
record if there are 00:00 hours or negative hours.

see - https://tinyurl.com/2okxbnpr
configuration - https://tinyurl.com/2gw5t2ex
sentry - 3978288761

closes odoo/enterprise#37872

Signed-off-by: Anh Thao PHAM <pta@odoo.com>