[FIX] l10n_mx_edi_pos: fix indeterministic bug in tour
In the tour `tour_invoice_previous_order`, we create, validate and pay
an order. Then, we go to the menu presenting the existing orders. In
this menu, we can filter the orders to get only the paid ones. Clicking
on one of these filters triggers a fetch to the server and the records
are sometimes loaded with a tiny delay.

This delay is causing an indeterminism: since the next step is to select
the first order in menu (`div.order-row:first`), we sometimes select the
current empty order, because the paid orders have not been loaded yet.
Adapting the trigger to select only the first *paid* order solves the
issue (`div.order-row:contains('Paid'):first`).

See for instance: https://runbot.odoo.com/runbot/build/24777641

closes odoo/enterprise#37881

Signed-off-by: Xavier Dollé (xdo) <xdo@odoo.com>