[IMP] knowledge: chatter and properties panel widgets
This commit adds 2 new widgets to the knowledge.article view:
the KnowledgeArticleChatter and the KnowledgeArticleProperties.

These widgets are OWL versions of the panels that were already present in the
original view meaning that their behavior is the same as before and the style
is similar.

For the KnowledgeArticleProperties, we are using the Component PropertiesField,
in order to modify or add properties to each article.

For the KnowledgeArticleChatter we are using the Chatter Component and we feed
it the ThreadModel `knowledge.article` and the threadId is dependant on the
article`s ID we also add all the necessary options for the Chatter to have the
same behavior as before.

The toggling of these components is handled via events on the bus sent by the
Topbar when the corresponding buttons are activated.

We are also removing the field `article_properties_is_empty` from the view and
the model because it is no longer needed, since we are checking directly via
the field `article_properties` if the properties are empty.

task-3096030

Part-of: odoo/enterprise#38074