[IMP] currency_rate_live: Added support to get the exchange rate from the SUNAT website instead of the BCRP
Currently, the service available in Odoo for the exchange rate in Peru is the Bank of Perú. The official rate in Peru is the one from the SUNAT. Bank of Peru is also official but actually the rates Bank of Peru uses, come from the SUNAT from this link https://www.sunat.gob.pe/a/txt/tipoCambio.txt. The problem is that the Bank of Peru currently being used in Odoo is down most of the time and creates more problems than it solves. The SUNAT link only brings the rate for the current date, which is a downside for the service but it is never down (it does not bring historic rates).

Solution:

We will implement the SUNAT service instead of the Bank of Peru and if customers need the historic rates, they can manually search on an official website (Bank of Peru, for instance) and import the rates needed.

The way this will be done is by substituting the current bank of Peru service with the SUNAT link in the same method and replacing the name with the SUNAT's. This will be done in stable since the Bank of Peru is causing more problems than it is solving.

closes odoo/enterprise#38884

X-original-commit: feda14f4e2df3e1a42059c2596be0218ff391f22
Signed-off-by: Josse Colpaert <jco@odoo.com>