[FIX] account_asset: deferred models creation fix
In the form view of a deferred expense/revenue, when quick
creating a model_id, it will actually create another deferred
asset.
It also set the new deferred asset as model_id, which is totally
wrong.

By adding the default_state to the context of the field, we can
ensure that it is a model that is going to be created.

This will also remove the quick create as it would lead to
unexpected results (model without accounts) and pass the
accounts of the asset in the context when creating a model
from it to improve the user experience.

Task id #3221592

closes odoo/enterprise#37893

Signed-off-by: John Laterre (jol) <jol@odoo.com>