[FIX] knowledge: fix "inverse" usage of is_user_favorite
Currently, the inverse method of the is_user_favorite field is implemented to
always toggle the value, no matter what we actually set on the record.

This commit fixes the method by correctly computing whether we want to enable
or disable the favorite.

The "action_toggle_favorite" method is adapted accordingly.
Some tests were also modified to ensure the behavior is consistent.

A query counter had to be modified as we need to invalidate the cache when
toggling the field, as the inverse will write on a separate model, it needs to
be re-evaluated completely to get the proper value.

Task-3207966

closes odoo/enterprise#37918

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>