[IMP] l10n_lu_reports: taxes test
We want to invigorate a Tax Name Taxonomy so that tax names are as codified as possible. This allow for a better display on Invoice Description, and allowed us to implement a smart name_search.

In the community PR we changed the name of the taxes, and those test are also impacted

closes odoo/enterprise#37911

Task-id: 3052677
Related: odoo/odoo#114597
Signed-off-by: John Laterre (jol) <jol@odoo.com>