[FIX] mrp: quantity not updated on tablet view
The quantity was not updated when entering the tablet view.
To recreate this behavior, you needed to have a workorder planned
on a workcenter that need authentication (everyone is allowed to work on it).

Then, no one should be logged and you need to click on "open tablet view"
in the list view.

Fixed by addind an additional parameter in the button_start method
that will by pass the connection checks.

task id : 3216277

Part-of: odoo/enterprise#38007