[FIX] helpdesk_timesheet: recompute total h. when changing timesheet h.
Current behaviour:
When updating the hours of a timesheet, it doesn't update the total
hours on the helpdesk ticket.

Expected behaviour:
The total hours should be updated when changing the timesheet
`unit_amount`, even outside of the Helpdesk app.

Steps to reproduce:
- Install Helpdesk, Timesheets, helpdesk_timesheet
- On a helpdesk ticket, log 0h30, add a description, the total hours
  is 0h30
- In Timesheets > List view > search the timesheet based on the
  description, change the hours to something else than 0h30
- Go back to the helpdesk ticket, notice the total hours hasn't updated.

Reason for the problem:
Lack of dependency on `timesheet_ids.unit_amount` for the compute of
the total hours on the ticket `_compute_total_hours_spent`.

Fix:
Add the missing dependency.

Affected versions:
- 15.0
- saas-15.2
- 16.0
- saas-16.1
- master

opw-3196050

closes odoo/enterprise#38185

X-original-commit: f924eab2a016cf8e3aece99acab523b8749e5852
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>